fix: renaming modal
